rpma: endpoint pre-implementation cleanup
